Fixup minor bugs around widgets, cleanup code
