+++ widgets/grid/widget.scss
@@ -1,4 +1,5 @@
 @import "@styles/mixins.scss";
+@import "@styles/expanded-tile.scss";
 @import "@templates/tags/tags.scss";
 
 #nosto-ugc-container {
@@ -69,24 +70,5 @@
 
 .hidden {
   display: none;
-}
-
-.expanded-tile-overlay {
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  inset: 0;
-  background-color: rgb(0 0 0 / 90%);
-  z-index: 10;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-@include breakpoint(tablet) {
-  .expanded-tile-overlay {
-    align-items: flex-start;
-  }
 }
