+++ widgets/masonry/widget.scss
@@ -1,5 +1,6 @@
 @import "@styles/mixins.scss";
 @import "@templates/tags/tags.scss";
+@import "@styles/expanded-tile.scss";
 
 #nosto-ugc-container {
 .grid {
@@ -102,14 +103,4 @@
 
 .hidden {
   display: none;
-}
-
-.expanded-tile-overlay {
-  position: fixed;
-  width: 100%;
-  height: 100%;
-  inset: 0;
-  background-color: rgb(0 0 0 / 90%);
-  z-index: 10;
-  cursor: pointer;
 }
