+++ widgets/styles/expanded-tile.scss
@@ -1,0 +1,21 @@
+
+@import "./mixins.scss";
+
+.expanded-tile-overlay {
+  position: fixed;
+  width: 100vw;
+  height: 100vh;
+  inset: 0;
+  background-color: rgb(0 0 0 / 90%);
+  z-index: 10;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+@include breakpoint(tablet) {
+  .expanded-tile-overlay {
+    align-items: flex-start;
+  }
+}
